Root based eslint + separate env rules for apps/packages
