+++ apps/rcf-examples/package.json
@@ -4,9 +4,7 @@
   "private": true,
   "scripts": {
     "build": "vite build",
-    "dev": "vite --host 0.0.0.0 --port 3002 --clearScreen false",
-    "lint": "eslint . --max-warnings=0",
-    "lint:fix": "eslint . --fix"
+    "dev": "vite --host 0.0.0.0 --port 3002 --clearScreen false"
   },
   "jest": {
     "preset": "@yas/jest/presets/node"
@@ -29,7 +27,6 @@
     "@types/react-dom": "^18.0.11",
     "@yas/build-tools": "workspace:*",
     "@yas/tsconfig": "workspace:*",
-    "eslint-config-yas": "workspace:*",
     "vite": "^4.3.9"
   }
 }
