+++ apps/web/package.json
@@ -5,8 +5,6 @@
   "scripts": {
     "build": "yas-env vite build",
     "dev": "yas-env vite --host 0.0.0.0 --port 3002 --clearScreen false",
-    "lint": "eslint . --max-warnings=0",
-    "lint:fix": "eslint . --fix",
     "test": "pnpm test:unit && pnpm test:e2e",
     "test:watch": "concurrently -n unit,e2e \"pnpm test:unit --watch\" \"pnpm test:e2e --ui\"",
     "test:unit": "jest",
@@ -36,7 +34,6 @@
     "@yas/playwright": "workspace:*",
     "@yas/tsconfig": "workspace:*",
     "concurrently": "^8.0.1",
-    "eslint-config-yas": "workspace:*",
     "vite": "^4.3.9"
   }
 }
