+++ apps/web/src/env.ts
@@ -1,6 +1,7 @@
-import { z, mode } from "@yas/zod";
+import { z, mode, truthy } from "@yas/zod";
 
 const schema = z.object({
+  CI: truthy.default(false),
   mode: mode.default("development"),
   apiUrl: z.string(),
 });
