+++ packages/react-imperative-hook/.eslintrc.js
@@ -1,5 +1,4 @@
 module.exports = {
-  root: true,
   extends: ["yas"],
   rules: {
     // Due to frequent use of generics in this library we need to disable these rules
