+++ packages/react-imperative-hook/package.json
@@ -28,8 +28,6 @@
   "scripts": {
     "build": "pnpm typecheck && tsup",
     "typecheck": "tsc --noEmit --project tsconfig.json",
-    "lint": "eslint . --max-warnings=0",
-    "lint:fix": "eslint . --fix",
     "test": "jest",
     "test:watch": "jest --watch",
     "release": "publicize-package-json -c \"pnpm publish --no-git-checks\" -d dist"
