--- packages/shiki-twoslash-loader/src/env.js
@@ -1,3 +1,0 @@
-module.exports.env = {
-  userSettingsPath: process.env.SHIKI_TWOSLASH_SETTINGS_PATH,
-};
