--- packages/yas-build-tools/.eslintrc.js
@@ -1,4 +1,0 @@
-module.exports = {
-  root: true,
-  extends: ["yas"],
-};
