+++ packages/yas-env/src/validEnvFiles.js
@@ -1,0 +1,1 @@
+module.exports = ["src/env.ts", "src/env.js"];
