+++ turbo.json
@@ -6,12 +6,6 @@
       "dependsOn": ["^build"],
       "outputs": ["dist/**", "build/**"]
     },
-    "lint": {
-      "cache": false
-    },
-    "lint:fix": {
-      "cache": false
-    },
     "test": {
       "dependsOn": ["^test"]
     },
